requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
youtube-transcript-api==0.6.2
sqlalchemy==2.0.30
faiss-cpu==1.8.0
langchain==0.2.1
langchain-community==0.2.1
sentence-transformers==2.7.0
python-dotenv==1.0.1
requests==2.32.2
jinja2==3.1.4
pydantic==2.7.1
yt-dlp==2024.4.9
